mrc-misc-fn-modules(extract-pp-proxy-move,mechanical_provable): move pp-proxy topk-size resolution to misc_utils module (cut+paste)
